Refactor command tool structure and fix related imports
Restructured the `commands_tool.py` into a dedicated `commands` module and updated imports accordingly. Adjusted tests to reflect these changes and ensured compatibility with the new module structure. Improved readability and maintainability of the affected code.
