+++ dev_kit_mcp_server/tools/__init__.py
@@ -2,7 +2,7 @@
 
 import importlib.util
 
-from .commands_tool import ExecMakeTarget
+from .commands import ExecMakeTarget
 from .file_sys.create import CreateDirOperation
 from .file_sys.edit import EditFileOperation
 from .file_sys.move import MoveDirOperation
