+++ dev_kit_mcp_server/tools/commands/__init__.py
@@ -1,0 +1,5 @@
+"""Commands module for the dev_kit_mcp_server package."""
+
+from dev_kit_mcp_server.tools.commands.make_target import ExecMakeTarget
+
+__all__ = ["ExecMakeTarget"]
